requirements.txt
python-telegram-bot==13.7
urllib3
telethon